requirements.txt
asgiref==3.4.1
backcall==0.2.0
charset-normalizer==2.0.5
colorama==0.4.4
coreschema==0.0.4
decorator==5.1.0
Django==3.2.7
django-filter==2.4.0
djangorestframework==3.12.4
idna==3.2
ipython==7.27.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.20
psycopg2-binary==2.8.6
Pygments==2.10.0
pytz==2021.1
simplejson==3.17.5
sqlparse==0.4.2
traitlets==5.1.0
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5